// types/article.ts
// types/article.ts

export interface Article {
    id: number;
    title: string;
    slug: string;
    excerpt: string;
    content: string;
    image: string;
    author: string;
    date: string;
    category: string;
    categoryName: string;
    metaTitle: string;
    metaDescription: string;
    metaKeywords: string;
  }